+++ Source/AccelByteUe4Sdk/Public/Core/AccelByteDefines.h
@@ -5,8 +5,11 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include "Runtime/Launch/Resources/Version.h"
 
 #if (ENGINE_MAJOR_VERSION==5 && ENGINE_MINOR_VERSION>=0)
+#include "Containers/Ticker.h"
+
 using FTickerAlias = FTSTicker;
 using FDelegateHandleAlias = FTSTicker::FDelegateHandle;
 #else