+++ Source/AccelByteUe4Sdk/Public/Models/AccelByteSessionModels.h
@@ -29,7 +29,8 @@ enum class EAccelByteV2SessionMemberStatus : uint8
 	LEFT,
 	DROPPED,
 	REJECTED,
-	KICKED
+	KICKED,
+	TIMEOUT
 };
 
 UENUM(BlueprintType)